+++ src/app/view-config-upgraders.js
@@ -310,3 +310,16 @@ export function upgradeFrom1_0_6(config) {
     version: '1.0.7',
   };
 }
+
+// Added in version 1.0.8:
+// - Support for multiple `dataset` coordination scopes and
+// dataset-specific coordination scope mappings for all
+// other coordination types.
+export function upgradeFrom1_0_7(config) {
+  const newConfig = cloneDeep(config);
+
+  return {
+    ...newConfig,
+    version: '1.0.8',
+  };
+}