+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.
 
 ### Changed
+- Use a hash table lookup instead of calling `indexOf` repeatedly for the heatmap component tiling.
 
 ## [1.1.19](https://www.npmjs.com/package/vitessce/v/1.1.19) - 2022-03-30
 
@@ -15,7 +16,6 @@
 - Upgrade Viv to 0.12.6 to fix shader compilation issue with interleaved RGB images
 - Fixed layer controller raster channel slider bug, related to [MUI slider issue](https://github.com/mui/material-ui/issues/20896).
 - Started to update the documentation to use the term "view" rather than "component".
-- Use a hash table lookup instead of calling `indexOf` repeatedly for the heatmap component tiling.
 
 ## [1.1.18](https://www.npmjs.com/package/vitessce/v/1.1.18) - 2022-02-14