+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 ### Added
 - Add a tutorial that describes how to deploy a Vitessce web app to GitHub pages.
 - Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.
+- Adds new view config schema version `1.0.8` to support multiple `dataset` coordination scopes and dataset-specific coordination scope mappings for all other coordination types
+```js
+datasets: [
+  { uid: 'my-query', ... },
+  { uid: 'some-atlas', ... },
+],
+coordinationSpace: {
+  dataset: {
+    REFERENCE: 'some-atlas',
+    QUERY: 'my-query',
+  },
+  embeddingType: {
+    common: 'UMAP',
+  },
+  embeddingZoom: {
+    refZoom: 2,
+    qryZoom: 4,
+  },
+  ...,
+},
+layout: [
+  {
+    component: 'qrComparisonScatterplot',
+    coordinationScopes: {
+      dataset: ['REFERENCE', 'QUERY'],
+      embeddingType: 'common',
+      embeddingZoom: { REFERENCE: 'refZoom', QUERY: 'qryZoom' },
+    },
+    x: 0, y: 0, w: 5, h: 12,
+  },
+  ...,
+],
+...
+```
 
 ### Changed
 - Use a hash table lookup instead of calling `indexOf` repeatedly for the heatmap component tiling.