+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 
-### Added
 - Adds support for plugin view types and plugin coordination types.
 
+### Added
+- Add a tutorial that describes how to deploy a Vitessce web app to GitHub pages.
+- Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.
+
 ### Changed
+- Use a hash table lookup instead of calling `indexOf` repeatedly for the heatmap component tiling.
 
 ## [1.1.19](https://www.npmjs.com/package/vitessce/v/1.1.19) - 2022-03-30
-
-
 
 ### Added
 - Added a roadmap page to the documentation.