+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 
 ### Added
+
+### Changed
+
+## [1.1.20](https://www.npmjs.com/package/vitessce/v/1.1.20) - 2022-04-21
+
+
+
+### Added
 - Add a tutorial that describes how to deploy a Vitessce web app to GitHub pages.
 - Support `var` alias for AnnData to display altenrative gene names via new `geneAlias` field.
 - Adds new view config schema version `1.0.8` to support multiple `dataset` coordination scopes and dataset-specific coordination scope mappings for all other coordination types